Refactor test to use the public API
